Refactor client into sub-files and fix test.
